+++ python/GafferTest/TweakPlugTest.py
@@ -64,7 +64,10 @@ class TweakPlugTest( GafferTest.TestCase ) :
 
 	def testCreateCounterpart( self ) :
 
-		p = Gaffer.TweakPlug( "test", 10.0, Gaffer.TweakPlug.Mode.Multiply )
+		p = Gaffer.TweakPlug(
+			name = "p", nameDefault = "test", enabledDefault = False, modeDefault = Gaffer.TweakPlug.Mode.Multiply,
+			valuePlug = Gaffer.IntPlug(), flags = Gaffer.Plug.Flags.Default | Gaffer.Plug.Flags.Dynamic
+		)
 		p2 = p.createCounterpart( "p2", Gaffer.Plug.Direction.In )
 
 		self.assertIsInstance( p2, Gaffer.TweakPlug )
@@ -73,6 +76,7 @@ class TweakPlugTest( GafferTest.TestCase ) :
 		self.assertEqual( p2.keys(), p.keys() )
 		for n in p2.keys() :
 			self.assertIsInstance( p2[n], p[n].__class__ )
+			self.assertEqual( p2[n].defaultValue(), p[n].defaultValue() )
 
 	def testTweakParameters( self ) :
 
@@ -133,10 +137,55 @@ class TweakPlugTest( GafferTest.TestCase ) :
 		self.assertEqual( p2.direction(), Gaffer.Plug.Direction.In )
 		self.assertEqual( p2.keys(), p.keys() )
 
-	def testOldSerialisation( self ) :
+	def testLoadFrom0_53( self ) :
 
-		# Old scripts call a constructor with an outdated signature as below.
+		# The ancient constructor as used by serialisations in Gaffer 0.53 and earlier.
 		plug = Gaffer.TweakPlug( "exposure", flags = Gaffer.Plug.Flags.Default | Gaffer.Plug.Flags.Dynamic )
+		# This left the plug in a bogus half-initialised state that needed to be
+		# finalised by a subsequent `addChild()` call that provided the `value`
+		# plug.
+		self.assertEqual( plug.getName(), "exposure" )
+		self.assertIsInstance( plug["name"], Gaffer.StringPlug )
+		self.assertEqual( plug["name"].defaultValue(), "" )
+		self.assertEqual( plug["name"].getValue(), "" )
+		self.assertIsInstance( plug["enabled"], Gaffer.BoolPlug )
+		self.assertEqual( plug["enabled"].defaultValue(), True )
+		self.assertEqual( plug["enabled"].getValue(), True )
+		self.assertIsInstance( plug["mode"], Gaffer.IntPlug )
+		self.assertEqual( plug["mode"].defaultValue(), plug.Mode.Replace )
+		self.assertEqual( plug["mode"].getValue(), plug.Mode.Replace )
+		self.assertNotIn( "value", plug )
+		self.assertEqual( plug.getFlags(), Gaffer.Plug.Flags.Default | Gaffer.Plug.Flags.Dynamic )
+
+	def testLoadFrom1_5_16_1( self ) :
+
+		script = Gaffer.ScriptNode()
+		script["fileName"].setValue( Gaffer.rootPath() / "python" / "GafferTest" / "scripts" / "tweakPlug-1.5.16.1.gfr" )
+		script.load()
+
+		def assertExpected( script ) :
+
+			plug = script["Node"]["user"]["tweak"]
+			self.assertIsInstance( plug, Gaffer.TweakPlug )
+			self.assertIsInstance( plug["name"], Gaffer.StringPlug )
+			self.assertEqual( plug["name"].defaultValue(), "" )
+			self.assertEqual( plug["name"].getValue(), "tweakName" )
+			self.assertIsInstance( plug["enabled"], Gaffer.BoolPlug )
+			self.assertEqual( plug["enabled"].defaultValue(), True )
+			self.assertEqual( plug["enabled"].getValue(), True )
+			self.assertIsInstance( plug["mode"], Gaffer.IntPlug )
+			self.assertEqual( plug["mode"].defaultValue(), plug.Mode.Replace )
+			self.assertEqual( plug["mode"].getValue(), plug.Mode.Add )
+			self.assertIsInstance( plug["value"], Gaffer.IntPlug )
+			self.assertEqual( plug["value"].defaultValue(), 1 )
+			self.assertEqual( plug["value"].getValue(), 2 )
+			self.assertEqual( plug.getFlags(), Gaffer.Plug.Flags.Default | Gaffer.Plug.Flags.Dynamic )
+
+		assertExpected( script )
+
+		script2 = Gaffer.ScriptNode()
+		script2.execute( script.serialise() )
+		assertExpected( script )
 
 	def testApplyReturnValues( self ) :
 
@@ -594,6 +643,56 @@ class TweakPlugTest( GafferTest.TestCase ) :
 
 		with GafferTest.TestRunner.PerformanceScope() :
 			tweaks.applyTweaks( parameters )
+
+	def testFullConstructor( self ) :
+
+		valuePlug = Gaffer.IntPlug()
+
+		plug = Gaffer.TweakPlug( "name", "tweakName", False, Gaffer.TweakPlug.Mode.Add, valuePlug, Gaffer.Plug.Flags.Default | Gaffer.Plug.Flags.Dynamic )
+		self.assertEqual( plug.getName(), "name" )
+		self.assertTrue( plug.isSetToDefault() )
+		self.assertEqual( plug["name"].getValue(), "tweakName" )
+		self.assertEqual( plug["enabled"].getValue(), False )
+		self.assertEqual( plug["mode"].getValue(), Gaffer.TweakPlug.Mode.Add )
+		self.assertEqual( plug["mode"].getValue(), Gaffer.TweakPlug.Mode.Add )
+		self.assertTrue( plug["value"].isSame( valuePlug ) )
+
+		# As above, but using keyword arguments.
+		plug = Gaffer.TweakPlug(
+			nameDefault = "tweakName",
+			name = "name",
+			enabledDefault = False,
+			modeDefault = Gaffer.TweakPlug.Mode.Add,
+			flags = Gaffer.Plug.Flags.Default | Gaffer.Plug.Flags.Dynamic,
+			valuePlug = valuePlug
+		)
+		self.assertEqual( plug.getName(), "name" )
+		self.assertTrue( plug.isSetToDefault() )
+		self.assertEqual( plug["name"].getValue(), "tweakName" )
+		self.assertEqual( plug["enabled"].getValue(), False )
+		self.assertEqual( plug["mode"].getValue(), Gaffer.TweakPlug.Mode.Add )
+		self.assertEqual( plug["mode"].getValue(), Gaffer.TweakPlug.Mode.Add )
+		self.assertTrue( plug["value"].isSame( valuePlug ) )
+
+	def testSerialiseChildDefaultValues( self ) :
+
+		script = Gaffer.ScriptNode()
+		script["node"] = Gaffer.Node()
+		script["node"]["user"]["tweak"] = Gaffer.TweakPlug(
+			nameDefault = "tweakName",
+			valuePlug = Gaffer.IntPlug( defaultValue = 10 ),
+			enabledDefault = False,
+			modeDefault = Gaffer.TweakPlug.Mode.Subtract,
+			flags = Gaffer.Plug.Flags.Default | Gaffer.Plug.Flags.Dynamic
+		)
+
+		script2 = Gaffer.ScriptNode()
+		script2.execute( script.serialise() )
+
+		self.assertEqual( script2["node"]["user"]["tweak"]["name"].defaultValue(), "tweakName" )
+		self.assertEqual( script2["node"]["user"]["tweak"]["enabled"].defaultValue(), False )
+		self.assertEqual( script2["node"]["user"]["tweak"]["mode"].defaultValue(), Gaffer.TweakPlug.Mode.Subtract )
+		self.assertEqual( script2["node"]["user"]["tweak"]["value"].defaultValue(), 10 )
 
 if __name__ == "__main__":
 	unittest.main()